Refactor JsonApi
