+++ elide-core/src/main/java/com/yahoo/elide/core/security/RequestScope.java
@@ -13,7 +13,25 @@ import com.yahoo.elide.jsonapi.document.processors.WithMetadata;
  * The request scope interface passed to checks.
  */
 public interface RequestScope extends WithMetadata {
+    /**
+     * Returns the {@link User} of the request.
+     *
+     * @return the user of the request
+     */
     User getUser();
+
+    /**
+     * Returns the {@link Route} of the request which contains the base url, path,
+     * headers, parameters and api version of the request.
+     *
+     * @return the route of the request
+     */
     Route getRoute();
+
+    /**
+     * Returns the {@link DataStoreTransaction} of the request.
+     *
+     * @return the transaction of the request
+     */
     DataStoreTransaction getTransaction();
 }
