+++ elide-integration-tests/src/test/java/com/yahoo/elide/assignedIdLongTests/AssignedIdLongIT.java
@@ -5,7 +5,6 @@
  */
 package com.yahoo.elide.assignedIdLongTests;
 
-import static com.yahoo.elide.Elide.JSONAPI_CONTENT_TYPE;
 import static com.yahoo.elide.test.jsonapi.JsonApiDSL.attr;
 import static com.yahoo.elide.test.jsonapi.JsonApiDSL.attributes;
 import static com.yahoo.elide.test.jsonapi.JsonApiDSL.datum;
@@ -16,6 +15,7 @@ import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.equalTo;
 
 import com.yahoo.elide.initialization.IntegrationTest;
+import com.yahoo.elide.jsonapi.JsonApi;
 import com.yahoo.elide.test.jsonapi.elements.Data;
 import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.Test;
@@ -45,8 +45,8 @@ public class AssignedIdLongIT extends IntegrationTest {
         );
 
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
-                .accept(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
+                .accept(JsonApi.MEDIA_TYPE)
                 .body(original)
                 .post("/assignedIdLong")
                 .then()
@@ -54,8 +54,8 @@ public class AssignedIdLongIT extends IntegrationTest {
                 .body(equalTo(original.toJSON()));
 
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
-                .accept(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
+                .accept(JsonApi.MEDIA_TYPE)
                 .body(modified)
                 .patch("/assignedIdLong/1")
                 .then()
