+++ elide-integration-tests/src/test/java/com/yahoo/elide/async/integration/tests/AsyncIT.java
@@ -5,7 +5,6 @@
  */
 package com.yahoo.elide.async.integration.tests;
 
-import static com.yahoo.elide.Elide.JSONAPI_CONTENT_TYPE;
 import static com.yahoo.elide.core.dictionary.EntityDictionary.NO_VERSION;
 import static com.yahoo.elide.test.graphql.GraphQLDSL.UNQUOTED_VALUE;
 import static com.yahoo.elide.test.graphql.GraphQLDSL.argument;
@@ -39,6 +38,7 @@ import com.yahoo.elide.core.dictionary.EntityDictionary;
 import com.yahoo.elide.core.exceptions.HttpStatus;
 import com.yahoo.elide.core.request.route.Route;
 import com.yahoo.elide.core.security.User;
+import com.yahoo.elide.jsonapi.JsonApi;
 import com.yahoo.elide.jsonapi.resources.SecurityContextUser;
 import com.yahoo.elide.test.graphql.EnumFieldSerializer;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -89,7 +89,7 @@ public class AsyncIT extends AsyncApiIT {
 
         //Create Async Request
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
                 .header("sleep", "1000")
                 .body(
                         data(
@@ -144,7 +144,7 @@ public class AsyncIT extends AsyncApiIT {
 
         //Create Async Request
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
                 .header("sleep", "1000")
                 .body(
                         data(
@@ -350,7 +350,7 @@ public class AsyncIT extends AsyncApiIT {
     @Test
     public void jsonApiUnknownRequestTests() throws InterruptedException {
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
                 .body(
                         data(
                                 resource(
@@ -432,7 +432,7 @@ public class AsyncIT extends AsyncApiIT {
     public void noReadEntityTests() throws InterruptedException {
         //Create Async Request
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
                 .body(
                         data(
                                 resource(
@@ -500,6 +500,7 @@ public class AsyncIT extends AsyncApiIT {
                         .withAuditLogger(new TestAuditLogger()).build());
 
         elide.doScans();
+        JsonApi jsonApi = new JsonApi(elide);
 
         User ownerUser = new User(() -> "owner-user");
         SecurityContextUser securityContextAdminUser = new SecurityContextUser(new SecurityContext() {
@@ -542,15 +543,15 @@ public class AsyncIT extends AsyncApiIT {
         String baseUrl = "/";
         // Principal is Owner
         Route route = Route.builder().baseUrl(baseUrl).path("/asyncQuery/" + id).apiVersion(NO_VERSION).build();
-        response = elide.get(route, ownerUser, null);
+        response = jsonApi.get(route, ownerUser, null);
         assertEquals(HttpStatus.SC_OK, response.getResponseCode());
 
         // Principal has Admin Role
-        response = elide.get(route, securityContextAdminUser, null);
+        response = jsonApi.get(route, securityContextAdminUser, null);
         assertEquals(HttpStatus.SC_OK, response.getResponseCode());
 
         // Principal without Admin Role
-        response = elide.get(route, securityContextNonAdminUser, null);
+        response = jsonApi.get(route, securityContextNonAdminUser, null);
         assertEquals(HttpStatus.SC_NOT_FOUND, response.getResponseCode());
     }
 
@@ -565,7 +566,7 @@ public class AsyncIT extends AsyncApiIT {
 
         //Create Async Request
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
                 .header("sleep", "1000")
                 .body(
                         data(
