+++ elide-integration-tests/src/test/java/com/yahoo/elide/async/integration/tests/TableExportIT.java
@@ -5,7 +5,6 @@
  */
 package com.yahoo.elide.async.integration.tests;
 
-import static com.yahoo.elide.Elide.JSONAPI_CONTENT_TYPE;
 import static com.yahoo.elide.core.dictionary.EntityDictionary.NO_VERSION;
 import static com.yahoo.elide.test.graphql.GraphQLDSL.UNQUOTED_VALUE;
 import static com.yahoo.elide.test.graphql.GraphQLDSL.argument;
@@ -42,6 +41,7 @@ import com.yahoo.elide.core.dictionary.EntityDictionary;
 import com.yahoo.elide.core.exceptions.HttpStatus;
 import com.yahoo.elide.core.request.route.Route;
 import com.yahoo.elide.core.security.User;
+import com.yahoo.elide.jsonapi.JsonApi;
 import com.yahoo.elide.jsonapi.resources.SecurityContextUser;
 import com.yahoo.elide.test.graphql.EnumFieldSerializer;
 import com.yahoo.elide.test.jsonapi.elements.Resource;
@@ -114,7 +114,7 @@ public class TableExportIT extends AsyncApiIT {
 
         //Create Table Export Request
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
                 .header("sleep", "1000")
                 .body(
                         data(
@@ -176,7 +176,7 @@ public class TableExportIT extends AsyncApiIT {
 
         //Create TableExport Request
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
                 .header("sleep", "1000")
                 .body(
                         data(
@@ -587,7 +587,7 @@ public class TableExportIT extends AsyncApiIT {
     @Test
     public void jsonApiUnknownRequestTests() throws InterruptedException {
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
                 .body(
                         data(
                                 resource(
@@ -632,7 +632,7 @@ public class TableExportIT extends AsyncApiIT {
     @Test
     public void jsonApiBadExportQueryTests() throws InterruptedException {
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
                 .body(
                         data(
                                 resource(
@@ -783,7 +783,7 @@ public class TableExportIT extends AsyncApiIT {
     public void jsonAPIRelationshipFetchTests() throws InterruptedException {
         //Create Table Export Request
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
                 .body(
                         data(
                                 resource(
@@ -1064,8 +1064,8 @@ public class TableExportIT extends AsyncApiIT {
         );
 
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
-                .accept(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
+                .accept(JsonApi.MEDIA_TYPE)
                 .body(
                         datum(noRead).toJSON()
                 )
@@ -1075,7 +1075,7 @@ public class TableExportIT extends AsyncApiIT {
 
         //Create Table Export Request
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
                 .body(
                         data(
                                 resource(
@@ -1146,7 +1146,7 @@ public class TableExportIT extends AsyncApiIT {
         Elide elide = new Elide(new ElideSettingsBuilder(dataStore)
                         .withEntityDictionary(dictionary)
                         .withAuditLogger(new TestAuditLogger()).build());
-
+        JsonApi jsonApi = new JsonApi(elide);
         User ownerUser = new User(() -> "owner-user");
         SecurityContextUser securityContextAdminUser = new SecurityContextUser(new SecurityContext() {
             @Override
@@ -1188,15 +1188,15 @@ public class TableExportIT extends AsyncApiIT {
         String baseUrl = "/";
         // Principal is Owner
         Route route = Route.builder().baseUrl(baseUrl).path("/tableExport/" + id).apiVersion(NO_VERSION).build();
-        response = elide.get(route, ownerUser, null);
+        response = jsonApi.get(route, ownerUser, null);
         assertEquals(HttpStatus.SC_OK, response.getResponseCode());
 
         // Principal has Admin Role
-        response = elide.get(route, securityContextAdminUser, null);
+        response = jsonApi.get(route, securityContextAdminUser, null);
         assertEquals(HttpStatus.SC_OK, response.getResponseCode());
 
         // Principal without Admin Role
-        response = elide.get(route, securityContextNonAdminUser, null);
+        response = jsonApi.get(route, securityContextNonAdminUser, null);
         assertEquals(HttpStatus.SC_NOT_FOUND, response.getResponseCode());
     }
 
@@ -1211,7 +1211,7 @@ public class TableExportIT extends AsyncApiIT {
 
         //Create Async Request
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
                 .header("sleep", "1000")
                 .body(
                         data(
