+++ elide-integration-tests/src/test/java/com/yahoo/elide/fieldLevelTest/FieldLevelParentClassIdIT.java
@@ -5,7 +5,6 @@
  */
 package com.yahoo.elide.fieldLevelTest;
 
-import static com.yahoo.elide.Elide.JSONAPI_CONTENT_TYPE;
 import static com.yahoo.elide.test.jsonapi.JsonApiDSL.attr;
 import static com.yahoo.elide.test.jsonapi.JsonApiDSL.attributes;
 import static com.yahoo.elide.test.jsonapi.JsonApiDSL.datum;
@@ -17,6 +16,7 @@ import static org.hamcrest.Matchers.equalTo;
 
 import com.yahoo.elide.core.utils.JsonParser;
 import com.yahoo.elide.initialization.IntegrationTest;
+import com.yahoo.elide.jsonapi.JsonApi;
 import com.yahoo.elide.test.jsonapi.elements.Resource;
 import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.Test;
@@ -45,8 +45,8 @@ public class FieldLevelParentClassIdIT extends IntegrationTest {
         );
 
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
-                .accept(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
+                .accept(JsonApi.MEDIA_TYPE)
                 .body(datum(original))
                 .post("/fieldLevelChild")
                 .then()
@@ -54,8 +54,8 @@ public class FieldLevelParentClassIdIT extends IntegrationTest {
                 .body(equalTo(datum(original).toJSON()));
 
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
-                .accept(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
+                .accept(JsonApi.MEDIA_TYPE)
                 .body(datum(modified))
                 .patch("/fieldLevelChild/1")
                 .then()
