+++ elide-standalone/src/test/java/example/ElideStandaloneConfigStoreTest.java
@@ -6,7 +6,6 @@
 
 package example;
 
-import static com.yahoo.elide.Elide.JSONAPI_CONTENT_TYPE;
 import static com.yahoo.elide.test.graphql.GraphQLDSL.argument;
 import static com.yahoo.elide.test.graphql.GraphQLDSL.arguments;
 import static com.yahoo.elide.test.graphql.GraphQLDSL.field;
@@ -33,6 +32,7 @@ import com.yahoo.elide.core.security.checks.prefab.Role;
 import com.yahoo.elide.core.type.Type;
 import com.yahoo.elide.core.utils.ClassScanner;
 import com.yahoo.elide.core.utils.coerce.CoerceUtil;
+import com.yahoo.elide.jsonapi.JsonApi;
 import com.yahoo.elide.modelconfig.DynamicConfiguration;
 import com.yahoo.elide.modelconfig.store.models.ConfigChecks;
 import com.yahoo.elide.standalone.ElideStandalone;
@@ -487,7 +487,7 @@ public class ElideStandaloneConfigStoreTest {
                 + "}";
 
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
                 .body(
                         datum(
                                 resource(
@@ -564,7 +564,7 @@ public class ElideStandaloneConfigStoreTest {
                 + "}";
 
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
                 .body(
                         datum(
                                 resource(
@@ -589,7 +589,7 @@ public class ElideStandaloneConfigStoreTest {
         String hjson = "#!/bin/sh ...";
 
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
                 .body(
                         datum(
                                 resource(
@@ -634,7 +634,7 @@ public class ElideStandaloneConfigStoreTest {
                 + "}";
 
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
                 .body(
                         datum(
                                 resource(
@@ -679,7 +679,7 @@ public class ElideStandaloneConfigStoreTest {
                 + "}";
 
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
                 .body(
                         datum(
                                 resource(
@@ -698,7 +698,7 @@ public class ElideStandaloneConfigStoreTest {
                 .statusCode(HttpStatus.SC_CREATED);
 
         given()
-                .contentType(JSONAPI_CONTENT_TYPE)
+                .contentType(JsonApi.MEDIA_TYPE)
                 .body(
                         datum(
                                 resource(
